Adicionando entrada por Scanner - @Renatu5

+++ src/main/java/com/example/project/bee1010/PackageData.java
@@ -3,8 +3,8 @@ package com.example.project.bee1010;
 public class PackageData {
     public int idProduct;
     public int quantity;
-    public double price;
-    public PackageData(int idProduct, int quantity, double price){
+    public float price;
+    public PackageData(int idProduct, int quantity, float price){
         this.idProduct = idProduct;
         this.quantity = quantity;
         this.price = price;
